allow fonts to share fontId

+++ minikin/library.h
@@ -41,18 +41,19 @@ extern "C" void destroy_font_collection(FontCollectionToken fontCollection);
 // glyphs (e.g. spaces), so just use a rough estimate in that case.
 //let space_advance = size.to_f32_px() / 3.0;
 
-extern "C" Layout* layout_text(const uint8_t* text,
+extern "C" Layout* layout_text(
+        const uint8_t* text,
         uint32_t textLength,
-                               bool isRtl,
-                               float_t fontSize,
-                               std::shared_ptr<FontCollection>* fontCollection);
+        bool isRtl,
+        float_t fontSize,
+        std::shared_ptr<FontCollection>* fontCollection);
 
 extern "C" void destroy_layout(Layout* layout);
 
 extern "C" MinikinRect get_bounds(const Layout* layout);
 extern "C" float_t get_advance(const Layout* layout);
 extern "C" size_t glyphs_count(const Layout* layout);
-extern "C" uint32_t get_font_id(const Layout* layout, size_t i);
+extern "C" int64_t get_font_id(const Layout* layout, size_t i);
 extern "C" uint32_t get_glyph_id(const Layout* layout, size_t i);
 extern "C" float_t get_x(const Layout* layout, size_t i);
 extern "C" float_t get_y(const Layout* layout, size_t i);
